backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-docx>=1.1.0
PyPDF2>=3.0.1
spacy>=3.7.2
scikit-learn>=1.4.0
numpy>=1.26.3
pandas>=2.2.0
openai>=1.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0
aiofiles>=23.2.1
Pillow>=10.2.0
reportlab>=4.0.8
jinja2>=3.1.3
python-dotenv>=1.0.0
